modularize: split one file into multiple sources
