+++ package.json
@@ -1,12 +1,14 @@
 {
   "name": "alps-unified",
   "version": "1.0.0",
-  "description": "CLI utility for ALPS translations",
+  "description": "ALPS translator, API and CLI",
   "main": "src/index.js",
   "bin": {
-    "unified": "src/index.js"
+    "unified": "bin/unified-cli.js"
   },
   "scripts": {
+    "build": "npm run build-webpack",
+    "build-webpack": "webpack",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -33,8 +35,16 @@
   "dependencies": {
     "boxen": "^4.0.0",
     "chalk": "^2.4.2",
+    "esm": "^3.2.25",
     "jstoxml": "^1.6.5",
     "yamljs": "^0.3.0",
     "yargs": "^13.2.4"
+  },
+  "engines": {
+    "node": ">=12"
+  },
+  "devDependencies": {
+    "webpack": "^4.46.0",
+    "webpack-cli": "^4.6.0"
   }
 }
