+++ src/translator/alps-json.js
@@ -1,0 +1,14 @@
+// ****************************************************
+// to ALPS JSON
+// ****************************************************
+
+module.exports = (doc, options) => {
+  var rtn = ""; 
+  try {
+    rtn = JSON.stringify(doc, null, 2);
+  }
+  catch(err) {
+    console.log(`ERROR: ${err}`);
+  }
+  return rtn
+}
