+++ src/translator/wsdl.js
@@ -1,0 +1,17 @@
+// ****************************************************
+// to WSDL
+// ****************************************************
+module.exports = (doc, options) => {
+  
+  var rtn = ""; 
+  
+  rtn += `<?xml version = '1.0' encoding = 'UTF-8'?>\n`;
+  rtn += `<!-- generated by "unified" from ${options.file} -->\n`;
+  rtn += `<!-- created: ${new Date()} -->\n`;
+  rtn += '<!-- source: http://github.com/mamund/2020-11-unified -->\n';
+  rtn += '\n';
+  rtn += "<definitions>\n";
+  rtn += "  <todo />\n";
+  rtn += "</definitions>\n";  
+  return rtn
+}
